MCS: Organize code a little
