+++ go.mod
@@ -12,10 +12,11 @@ require (
 	github.com/hashicorp/vault/api v1.0.4
 	github.com/lib/pq v1.2.0 // indirect
 	github.com/minio/cli v1.22.0
-	github.com/minio/minio v0.0.0-20200311155636-0af62d35a0c8
+	github.com/minio/minio v0.0.0-20200312144740-ed4bd20a7cfc
 	github.com/stretchr/testify v1.5.1 // indirect
+	golang.org/x/crypto v0.0.0-20200214034016-1d94cc7ab1c6 // indirect
 	golang.org/x/net v0.0.0-20200301022130-244492dfa37a // indirect
-	golang.org/x/sys v0.0.0-20191029155521-f43be2a4598c // indirect
+	golang.org/x/sys v0.0.0-20200212091648-12a6c2dcc1e4 // indirect
 	google.golang.org/genproto v0.0.0-20191115221424-83cc0476cb11 // indirect
 	google.golang.org/grpc v1.24.0 // indirect
 	k8s.io/api v0.17.3
